refactor: 공백, import 정리, 패키지 변경
