+++ src/main/java/blackjack/model/participant/Participant.java
@@ -1,6 +1,5 @@
 package blackjack.model.participant;
 
-import blackjack.model.Name;
 import blackjack.model.card.Card;
 import blackjack.model.card.CardDeck;
 import blackjack.model.card.CardScore;
