+++ src/main/java/blackjack/model/state/InitialState.java
@@ -1,6 +1,6 @@
 package blackjack.model.state;
 
-import blackjack.Hand;
+import blackjack.model.participant.Hand;
 import blackjack.model.card.Card;
 import blackjack.model.card.CardDeck;
