+++ src/main/java/blackjack/model/state/State.java
@@ -1,6 +1,6 @@
 package blackjack.model.state;
 
-import blackjack.Hand;
+import blackjack.model.participant.Hand;
 import blackjack.model.card.Card;
 import blackjack.model.card.CardDeck;
 import blackjack.model.card.CardScore;
