+++ src/test/java/blackjack/model/NameTest.java
@@ -1,5 +1,6 @@
 package blackjack.model;
 
+import blackjack.model.participant.Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
